Deleted scan results, added copyright, moved ManagerFields from internal
