+++ galasa-managers-parent/galasa-managers-cicsts-parent/dev.galasa.cicsts.ceda.manager.ivt/src/main/java/dev/galasa/cicsts/ceda/manager/ivt/CEDAManagerIVT.java
@@ -1,3 +1,8 @@
+/*
+ * Licensed Materials - Property of IBM
+ * 
+ * (c) Copyright IBM Corp. 2020.
+ */
 package dev.galasa.cicsts.ceda.manager.ivt;
 
 import static org.assertj.core.api.Assertions.*;
