+++ galasa-managers-parent/galasa-managers-cicsts-parent/dev.galasa.cicsts.ceda.manager/src/main/java/dev/galasa/cicsts/ceda/CEDA.java
@@ -1,11 +1,14 @@
+/*
+ * Licensed Materials - Property of IBM
+ * 
+ * (c) Copyright IBM Corp. 2020.
+ */
 package dev.galasa.cicsts.ceda;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-
-import dev.galasa.cicsts.ceda.internal.CEDAManagerField;
 import dev.galasa.framework.spi.ValidAnnotatedFields;
 import dev.galasa.zos3270.ITerminal;
