+++ galasa-managers-parent/galasa-managers-cicsts-parent/dev.galasa.cicsts.cemt.manager/src/main/java/dev/galasa/cicsts/cemt/CEMT.java
@@ -1,3 +1,8 @@
+/*
+ * Licensed Materials - Property of IBM
+ * 
+ * (c) Copyright IBM Corp. 2020.
+ */
 package dev.galasa.cicsts.cemt;
 
 import java.lang.annotation.ElementType;
@@ -5,7 +10,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import dev.galasa.cicsts.cemt.internal.CEMTManagerField;
+import dev.galasa.cicsts.cemt.CEMTManagerField;
 import dev.galasa.framework.spi.ValidAnnotatedFields;
 
 @Retention(RetentionPolicy.RUNTIME)
