--- galasa-managers-parent/galasa-managers-parent_scan-results/License-Report.html
@@ -1,1 +1,0 @@
-<style>h2 strong{color: red;} pre{padding-left:1em; background-color:cornsilk; border:1px solid black; margin: 0 2em}</style><h1 id="nolicenseswerefoundwhichrequirereview">No licenses were found which require review.</h1>
