--- galasa-managers-parent/galasa-managers-parent_scan-results/License-Report.md
@@ -1,1 +1,0 @@
-# No licenses were found which require review.
